split code from dashboardpage into separate compnents and modals. fix some apis
