+++ src/App.jsx
@@ -5,11 +5,12 @@ import {
   Routes,
   Route,
   Navigate,
+  useLocation,
 } from "react-router-dom";
 import { AuthProvider } from "./context/AuthContext";
 import { ToastProvider } from "./context/ToastContext";
 import Navbar from "./components/Navbar";
-import DashboardLayout from "./components/DashboardLayout"; // NEW
+import DashboardLayout from "./components/DashboardLayout";
 import Toast from "./components/Toast";
 import LandingPage from "./pages/LandingPage";
 import DashboardPage from "./pages/DashboardPage";
@@ -23,7 +24,7 @@ import ResetPasswordPage from "./pages/ResetPasswordPage";
 import PrivateRoute from "./components/PrivateRoute";
 import BusinessPublicPage from "./pages/BusinessPublicPage";
 import EmailTemplatesPage from "./pages/EmailTemplatesPage";
-import CampaignManagementPage from "./pages/CampaignManagementPage"; // NEW
+import CampaignManagementPage from "./pages/CampaignManagementPage";
 import {
   NotFoundPage,
   ServerErrorPage,
@@ -43,27 +44,11 @@ import SmsSignupPage from "./pages/SmsSignupPage";
 import TwilioProofPage from "./pages/TwilioProofPage";
 import AutomationPage from "./pages/AutomationPage";
 
-function App() {
-  useEffect(() => {
-    document.body.style.overflow = '';
-  }, []);
-  const [isMaintenanceMode, setIsMaintenanceMode] = useState(false);
-  const [isCheckingMaintenance, setIsCheckingMaintenance] = useState(true);
-  
+// AppContent component - must be inside Router to use useLocation
+function AppContent() {
+  const location = useLocation();
 
-  // Check for maintenance mode on app load
-  useEffect(() => {
-    const checkMaintenanceMode = () => {
-      const envMaintenanceMode = import.meta.env.VITE_MAINTENANCE_MODE === "true";
-      const localMaintenanceMode = localStorage.getItem("maintenanceMode") === "true";
-      setIsMaintenanceMode(envMaintenanceMode || localMaintenanceMode);
-      setIsCheckingMaintenance(false);
-    };
-
-    checkMaintenanceMode();
-  }, []);
-
-  // Don't show navbar for certain routes
+  // Don't show navbar for authenticated routes
   const noNavbarRoutes = [
     "/login",
     "/register",
@@ -78,12 +63,169 @@ function App() {
     "/profile",
     "/automation",
     "/campaigns",
-    "/account/billing",
+    "/account",
   ];
 
   const shouldShowNavbar = !noNavbarRoutes.some((route) =>
-    window.location.pathname.startsWith(route)
+    location.pathname.startsWith(route)
   );
+
+  return (
+    <div className="App">
+      {shouldShowNavbar && <Navbar />}
+      <Toast />
+      <Routes>
+        {/* Public routes */}
+        <Route path="/" element={<LandingPage />} />
+        <Route path="/login" element={<LoginPage />} />
+        <Route path="/register" element={<RegisterPage />} />
+        <Route path="/forgot-password" element={<ForgotPasswordPage />} />
+        <Route path="/reset-password" element={<ResetPasswordPage />} />
+        <Route path="/business/:id" element={<BusinessPublicPage />} />
+        <Route path="/sms-signup" element={<SmsSignupPage />} />
+        <Route path="/sms-signup/:businessId" element={<SmsSignupPage />} />
+        <Route path="/twilio-proof" element={<TwilioProofPage />} />
+        <Route path="/pricing" element={<PricingPage />} />
+        <Route path="/checkout/success" element={<CheckoutPages.Success />} />
+        <Route path="/checkout/error" element={<CheckoutPages.Error />} />
+        <Route path="/feedback-gate/:customerId" element={<FeedbackGatePage />} />
+        <Route path="/feedback/:customerId" element={<CustomerFeedbackPage />} />
+        <Route path="/opt-in-policy" element={<OptInPolicy />} />
+
+        {/* Error routes */}
+        <Route path="/forbidden" element={<ForbiddenPage />} />
+        <Route path="/server-error" element={<ServerErrorPage />} />
+        <Route path="/maintenance" element={<MaintenancePage />} />
+
+        {/* Protected routes WITH SIDEBAR LAYOUT */}
+        <Route
+          path="/dashboard"
+          element={
+            <PrivateRoute>
+              <DashboardLayout>
+                <DashboardPage />
+              </DashboardLayout>
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path="/customers"
+          element={
+            <PrivateRoute>
+              <DashboardLayout>
+                <CustomerManagementPage />
+              </DashboardLayout>
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path="/contacts"
+          element={
+            <PrivateRoute>
+              <DashboardLayout>
+                <ContactsPage />
+              </DashboardLayout>
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path="/email-templates"
+          element={
+            <PrivateRoute>
+              <DashboardLayout>
+                <EmailTemplatesPage />
+              </DashboardLayout>
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path="/review-requests"
+          element={
+            <PrivateRoute>
+              <DashboardLayout>
+                <ReviewRequestPage />
+              </DashboardLayout>
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path="/review-platform-setup"
+          element={
+            <PrivateRoute>
+              <DashboardLayout>
+                <ReviewPlatformSetupPage />
+              </DashboardLayout>
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path="/automation"
+          element={
+            <PrivateRoute>
+              <DashboardLayout>
+                <AutomationPage />
+              </DashboardLayout>
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path="/campaigns"
+          element={
+            <PrivateRoute>
+              <DashboardLayout>
+                <CampaignManagementPage />
+              </DashboardLayout>
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path="/profile"
+          element={
+            <PrivateRoute>
+              <DashboardLayout>
+                <ProfilePage />
+              </DashboardLayout>
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path="/account/billing"
+          element={
+            <PrivateRoute>
+              <DashboardLayout>
+                <AccountBillingPage />
+              </DashboardLayout>
+            </PrivateRoute>
+          }
+        />
+        <Route path="/account" element={<Navigate to="/account/billing" replace />} />
+
+        {/* 404 - Keep this LAST */}
+        <Route path="*" element={<NotFoundPage />} />
+      </Routes>
+    </div>
+  );
+}
+
+function App() {
+  useEffect(() => {
+    document.body.style.overflow = "";
+  }, []);
+
+  const [isMaintenanceMode, setIsMaintenanceMode] = useState(false);
+  const [isCheckingMaintenance, setIsCheckingMaintenance] = useState(true);
+
+  // Check for maintenance mode on app load
+  useEffect(() => {
+    const checkMaintenanceMode = () => {
+      const envMaintenanceMode = import.meta.env.VITE_MAINTENANCE_MODE === "true";
+      const localMaintenanceMode = localStorage.getItem("maintenanceMode") === "true";
+      setIsMaintenanceMode(envMaintenanceMode || localMaintenanceMode);
+      setIsCheckingMaintenance(false);
+    };
+
+    checkMaintenanceMode();
+  }, []);
 
   if (isCheckingMaintenance) {
     return (
@@ -109,139 +251,7 @@ function App() {
       <AuthProvider>
         <ToastProvider>
           <ErrorBoundary>
-            <div className="App">
-              {shouldShowNavbar && <Navbar />}
-              <Toast />
-              <Routes>
-                {/* Public routes */}
-                <Route path="/" element={<LandingPage />} />
-                <Route path="/login" element={<LoginPage />} />
-                <Route path="/register" element={<RegisterPage />} />
-                <Route path="/forgot-password" element={<ForgotPasswordPage />} />
-                <Route path="/reset-password" element={<ResetPasswordPage />} />
-                <Route path="/business/:id" element={<BusinessPublicPage />} />
-                <Route path="/sms-signup" element={<SmsSignupPage />} />
-                <Route path="/sms-signup/:businessId" element={<SmsSignupPage />} />
-                <Route path="/twilio-proof" element={<TwilioProofPage />} />
-                <Route path="/pricing" element={<PricingPage />} />
-                <Route path="/checkout/success" element={<CheckoutPages.Success />} />
-                <Route path="/checkout/error" element={<CheckoutPages.Error />} />
-                <Route path="/feedback-gate/:customerId" element={<FeedbackGatePage />} />
-                <Route path="/feedback/:customerId" element={<CustomerFeedbackPage />} />
-                <Route path="/opt-in-policy" element={<OptInPolicy />} />
-
-                {/* Error routes */}
-                <Route path="/forbidden" element={<ForbiddenPage />} />
-                <Route path="/server-error" element={<ServerErrorPage />} />
-                <Route path="/maintenance" element={<MaintenancePage />} />
-
-                {/* Protected routes WITH SIDEBAR LAYOUT */}
-                <Route
-                  path="/dashboard"
-                  element={
-                    <PrivateRoute>
-                      <DashboardLayout>
-                        <DashboardPage />
-                      </DashboardLayout>
-                    </PrivateRoute>
-                  }
-                />
-                <Route
-                  path="/customers"
-                  element={
-                    <PrivateRoute>
-                      <DashboardLayout>
-                        <CustomerManagementPage />
-                      </DashboardLayout>
-                    </PrivateRoute>
-                  }
-                />
-                <Route
-                  path="/contacts"
-                  element={
-                    <PrivateRoute>
-                      <DashboardLayout>
-                        <ContactsPage />
-                      </DashboardLayout>
-                    </PrivateRoute>
-                  }
-                />
-                <Route
-                  path="/email-templates"
-                  element={
-                    <PrivateRoute>
-                      <DashboardLayout>
-                        <EmailTemplatesPage />
-                      </DashboardLayout>
-                    </PrivateRoute>
-                  }
-                />
-                <Route
-                  path="/review-requests"
-                  element={
-                    <PrivateRoute>
-                      <DashboardLayout>
-                        <ReviewRequestPage />
-                      </DashboardLayout>
-                    </PrivateRoute>
-                  }
-                />
-                <Route
-                  path="/review-platform-setup"
-                  element={
-                    <PrivateRoute>
-                      <DashboardLayout>
-                        <ReviewPlatformSetupPage />
-                      </DashboardLayout>
-                    </PrivateRoute>
-                  }
-                />
-                <Route
-                  path="/automation"
-                  element={
-                    <PrivateRoute>
-                      <DashboardLayout>
-                        <AutomationPage />
-                      </DashboardLayout>
-                    </PrivateRoute>
-                  }
-                />
-                <Route
-                  path="/campaigns"
-                  element={
-                    <PrivateRoute>
-                      <DashboardLayout>
-                        <CampaignManagementPage />
-                      </DashboardLayout>
-                    </PrivateRoute>
-                  }
-                />
-                <Route
-                  path="/profile"
-                  element={
-                    <PrivateRoute>
-                      <DashboardLayout>
-                        <ProfilePage />
-                      </DashboardLayout>
-                    </PrivateRoute>
-                  }
-                />
-                <Route
-                  path="/account/billing"
-                  element={
-                    <PrivateRoute>
-                      <DashboardLayout>
-                        <AccountBillingPage />
-                      </DashboardLayout>
-                    </PrivateRoute>
-                  }
-                />
-                <Route path="/account" element={<Navigate to="/account/billing" replace />} />
-
-                {/* 404 - Keep this LAST */}
-                <Route path="*" element={<NotFoundPage />} />
-              </Routes>
-            </div>
+            <AppContent />
           </ErrorBoundary>
         </ToastProvider>
       </AuthProvider>
